jupyter api allows multiple integration files now
