+++ kotlin-spark-api/3.2/src/main/kotlin/org/jetbrains/kotlinx/spark/api/jupyter/SparkIntegration.kt
@@ -33,6 +33,7 @@ import java.io.InputStreamReader
 
 
 import org.apache.spark.*
+import org.jetbrains.kotlinx.jupyter.api.KotlinKernelHost
 import scala.collection.*
 import org.jetbrains.kotlinx.spark.api.SparkSession
 import scala.Product
@@ -40,62 +41,19 @@ import java.io.Serializable
 import scala.collection.Iterable as ScalaIterable
 import scala.collection.Iterator as ScalaIterator
 
+/**
+ * %use kotlin-spark-api
+ */
 @Suppress("UNUSED_VARIABLE", "LocalVariableName")
 @OptIn(ExperimentalStdlibApi::class)
-internal class SparkIntegration : JupyterIntegration() {
+internal class SparkIntegration : Integration() {
 
-    private val kotlinVersion = "1.6.21"
-    private val scalaCompatVersion = "2.12"
-    private val scalaVersion = "2.12.15"
-    private val spark3Version = "3.2.1"
+    override fun KotlinKernelHost.onLoaded() {
+        val _0 = execute("""%dumpClassesForSpark""")
 
-    override fun Builder.onLoaded() {
-
-        dependencies(
-            "org.apache.spark:spark-repl_$scalaCompatVersion:$spark3Version",
-            "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlinVersion",
-            "org.jetbrains.kotlin:kotlin-reflect:$kotlinVersion",
-            "org.apache.spark:spark-sql_$scalaCompatVersion:$spark3Version",
-            "org.apache.spark:spark-streaming_$scalaCompatVersion:$spark3Version",
-            "org.apache.spark:spark-mllib_$scalaCompatVersion:$spark3Version",
-            "org.apache.spark:spark-sql_$scalaCompatVersion:$spark3Version",
-            "org.apache.spark:spark-graphx_$scalaCompatVersion:$spark3Version",
-            "org.apache.spark:spark-launcher_$scalaCompatVersion:$spark3Version",
-            "org.apache.spark:spark-catalyst_$scalaCompatVersion:$spark3Version",
-            "org.apache.spark:spark-streaming_$scalaCompatVersion:$spark3Version",
-            "org.apache.spark:spark-core_$scalaCompatVersion:$spark3Version",
-            "org.scala-lang:scala-library:$scalaVersion",
-            "org.scala-lang.modules:scala-xml_$scalaCompatVersion:2.0.1",
-            "org.scala-lang:scala-reflect:$scalaVersion",
-            "org.scala-lang:scala-compiler:$scalaVersion",
-            "commons-io:commons-io:2.11.0",
-        )
-
-        println("SparkIntegration loaded")
-
-        import(
-            "org.jetbrains.kotlinx.spark.api.*",
-            "org.jetbrains.kotlinx.spark.api.tuples.*",
-            *(1..22).map { "scala.Tuple$it" }.toTypedArray(),
-            "org.apache.spark.sql.functions.*",
-            "org.apache.spark.*",
-            "org.apache.spark.sql.*",
-            "org.apache.spark.api.java.*",
-            "scala.collection.Seq",
-            "org.apache.spark.rdd.*",
-            "java.io.Serializable",
-            "org.apache.spark.streaming.api.java.*",
-            "org.apache.spark.streaming.api.*",
-            "org.apache.spark.streaming.*",
-        )
-
-        // onLoaded is only done for the non-streaming variant of kotlin-spark-api in the json file
-        onLoaded {
-            val _0 = execute("""%dumpClassesForSpark""")
-
-            @Language("kts")
-            val _1 = listOf(
-                """
+        @Language("kts")
+        val _1 = listOf(
+            """
                 val spark = org.jetbrains.kotlinx.spark.api.SparkSession
                     .builder()
                     .master(SparkConf().get("spark.master", "local[*]"))
@@ -103,52 +61,30 @@ internal class SparkIntegration : JupyterIntegration() {
                     .config("spark.sql.codegen.wholeStage", false)
                     .config("spark.io.compression.codec", "snappy")
                     .getOrCreate()""".trimIndent(),
-                """
+            """
                 spark.sparkContext.setLogLevel(org.jetbrains.kotlinx.spark.api.SparkLogLevel.ERROR)""".trimIndent(),
-                """
+            """
                 val sc by lazy { 
                     org.apache.spark.api.java.JavaSparkContext(spark.sparkContext) 
                 }""".trimIndent(),
-                """
+            """
                 println("Spark session has been started and is running. No `withSpark { }` necessary, you can access `spark` and `sc` directly. To use Spark streaming, use `%use kotlin-spark-api-streaming` instead.")""".trimIndent(),
-                """
+            """
                 inline fun <reified T> List<T>.toDS(): Dataset<T> = toDS(spark)""".trimIndent(),
-                """
+            """
                 inline fun <reified T> Array<T>.toDS(): Dataset<T> = spark.dsOf(*this)""".trimIndent(),
-                """
+            """
                 inline fun <reified T> dsOf(vararg arg: T): Dataset<T> = spark.dsOf(*arg)""".trimIndent(),
-                """
+            """
                 inline fun <reified T> RDD<T>.toDS(): Dataset<T> = toDS(spark)""".trimIndent(),
-                """
+            """
                 inline fun <reified T> JavaRDDLike<T, *>.toDS(): Dataset<T> = toDS(spark)""".trimIndent(),
-                """
+            """
                 inline fun <reified T> RDD<T>.toDF(): Dataset<Row> = toDF(spark)""".trimIndent(),
-                """
+            """
                 inline fun <reified T> JavaRDDLike<T, *>.toDF(): Dataset<Row> = toDF(spark)""".trimIndent(),
-                """
+            """
                 val udf: UDFRegistration get() = spark.udf()""".trimIndent(),
-            ).map(::execute)
-        }
-
-        onShutdown {
-            @Language("kts")
-            val _0 = execute("""
-                spark.stop()""".trimIndent()
-            )
-        }
-
-
-        // Render Dataset
-        render<Dataset<*>> {
-            HTML(it.toHtml())
-        }
-
-        render<RDD<*>> {
-            HTML(it.toJavaRDD().toHtml())
-        }
-
-        render<JavaRDDLike<*, *>> {
-            HTML(it.toHtml())
-        }
+        ).map(::execute)
     }
 }
