+++ kotlin-spark-api/3.2/src/main/resources/META-INF/kotlin-jupyter-libraries/libraries.json
@@ -3,6 +3,9 @@
   "producers": [
     {
       "fqn": "org.jetbrains.kotlinx.spark.api.jupyter.SparkIntegration"
+    },
+    {
+      "fqn": "org.jetbrains.kotlinx.spark.api.jupyter.SparkStreamingIntegration"
     }
   ]
 }
