+++ pom.xml
@@ -17,7 +17,7 @@
         <kotest-extensions-allure.version>1.1.0</kotest-extensions-allure.version>
         <embedded-kafka.version>3.1.0</embedded-kafka.version>
         <spark3.version>3.2.1</spark3.version>
-        <kotlin-jupyter-api.version>0.11.0-77</kotlin-jupyter-api.version>
+        <kotlin-jupyter-api.version>0.11.0-79</kotlin-jupyter-api.version>
         <kotlinx.html.version>0.7.3</kotlinx.html.version>
         <hadoop.version>3.3.1</hadoop.version>
 <!--        <junit.version>5.8.2</junit.version>-->
